fineshed dashboard-result-exams and corrections styles

+++ database/migrations/2025_10_21_001230_create_doctor_table.php
@@ -11,21 +11,23 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('doctor', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->foreignId('clinic_id')->constrained('clinic')->onDelete('cascade');
+        if (! Schema::hasTable('doctor')) {
+            Schema::create('doctor', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->foreignId('clinic_id')->constrained('clinic')->onDelete('cascade');
 
-            $table->string('crm');
-            $table->string('specialty');
-            $table->string('description')->nullable();
-            $table->decimal('average_rating')->nullable();
-            $table->decimal('consultation_fee');
-            $table->date('work_start_time');
-            $table->date('work_end_time');
-            $table->string('available_days');
-            $table->timestamps();
-        });
+                $table->string('crm');
+                $table->string('specialty');
+                $table->string('description')->nullable();
+                $table->decimal('average_rating')->nullable();
+                $table->decimal('consultation_fee');
+                $table->date('work_start_time');
+                $table->date('work_end_time');
+                $table->string('available_days');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
